make markdown async
