+++ ui/src/components/logs/LogLine.vue
@@ -37,7 +37,7 @@
 <script>
     import Convert from "ansi-to-html";
     import xss from "xss";
-    import Markdown from "../../utils/markdownSync";
+    import * as Markdown from "../../utils/markdown";
     import VRuntimeTemplate from "vue3-runtime-template";
     import MenuRight from "vue-material-design-icons/MenuRight.vue";
